requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
coverage==7.5.1
gunicorn==22.0.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
parse==1.20.1
pluggy==1.5.0
pytest==8.2.0
pytest-cov==5.0.0
requests==2.31.0
requests-wsgi-adapter==0.4.1
urllib3==2.2.1
WebOb==1.8.7
whitenoise==6.6.0
